Release 0.12.0

+++ src/main/java/ml/shifu/shifu/fs/ShifuFileUtils.java
@@ -18,23 +18,18 @@ package ml.shifu.shifu.fs;
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Scanner;
 import java.util.zip.GZIPInputStream;
-
-import ml.shifu.shifu.container.obj.ColumnConfig;
-import ml.shifu.shifu.container.obj.EvalConfig;
-import ml.shifu.shifu.container.obj.RawSourceData.SourceType;
-import ml.shifu.shifu.util.CommonUtils;
-import ml.shifu.shifu.util.Constants;
-import ml.shifu.shifu.util.HDFSUtils;
 
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorInputStream;
 import org.apache.commons.io.IOUtils;
@@ -48,9 +43,17 @@ import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.PathFilter;
 import org.apache.hadoop.io.compress.CompressionCodec;
 import org.apache.hadoop.io.compress.CompressionCodecFactory;
+import org.apache.hadoop.io.compress.SnappyCodec;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.xerial.snappy.SnappyInputStream;
+
+import ml.shifu.shifu.container.obj.ColumnConfig;
+import ml.shifu.shifu.container.obj.EvalConfig;
+import ml.shifu.shifu.container.obj.RawSourceData.SourceType;
+import ml.shifu.shifu.util.CommonUtils;
+import ml.shifu.shifu.util.Constants;
+import ml.shifu.shifu.util.HDFSUtils;
+import ml.shifu.shifu.util.HdfsPartFile;
 
 /**
  * ShifuFileUtils class encapsulate the file system interface from other components.
@@ -193,6 +196,30 @@ public class ShifuFileUtils {
         }
     }
 
+    /**
+     * Get InputStream from (Path, SourceType)
+     * @param path - file path
+     * @param sourceType - file type
+     * @return InputStream of file(Path, SourceType)
+     * @throws IOException - if fail to open file
+     */
+    public static InputStream getInputStream(Path path, SourceType sourceType) throws IOException {
+        try {
+            return getCompressInputStream(getFileSystemBySourceType(sourceType).open(path), path);
+        } catch (IOException e) {
+            // To manual fix a issue that FileSystem is closed exceptionally. Here we renew a FileSystem object to make
+            // sure all go through such issues.
+            if(e.getMessage() != null) {
+                if(e.getMessage().toLowerCase().indexOf("filesystem closed") >= 0) {
+                    if(sourceType == SourceType.HDFS) {
+                        return HDFSUtils.renewFS().open(path);
+                    }
+                }
+            }
+            throw e;
+        }
+    }
+
     private static InputStream getCompressInputStream(FSDataInputStream fdis, Path path) throws IOException {
         String name = path.getName();
         if(name.toLowerCase().endsWith(".gz")) {
@@ -200,7 +227,10 @@ public class ShifuFileUtils {
         } else if(name.toLowerCase().endsWith(".bz2")) {
             return new BZip2CompressorInputStream(fdis);
         } else if(name.toLowerCase().endsWith(".snappy")) {
-            return new SnappyInputStream(fdis);
+            Configuration conf = new Configuration();
+            CompressionCodecFactory ccf = new CompressionCodecFactory(conf);
+            CompressionCodec codec = ccf.getCodecByClassName(SnappyCodec.class.getName());
+            return codec.createInputStream(fdis);
         } else {
             return fdis;
         }
@@ -359,6 +389,33 @@ public class ShifuFileUtils {
     }
 
     /**
+     * Move src file to dst file in the same FileSystem.
+     * 
+     * @param srcPath
+     *            - source file to copy
+     * @param destPath
+     *            - destination file
+     * @param sourceType
+     *            - local/hdfs
+     * @throws IOException
+     *             - if any I/O exception in processing
+     */
+    public static void moveTo(String srcPath, String destPath, SourceType sourceType) throws IOException {
+        if(StringUtils.isEmpty(srcPath) || StringUtils.isEmpty(destPath) || sourceType == null) {
+            throw new IllegalArgumentException(String.format(
+                    "Null or empty parameters srcDataPath:%s, dstDataPath:%s, sourceType:%s", srcPath, destPath,
+                    sourceType));
+        }
+
+        FileSystem fs = getFileSystemBySourceType(sourceType);
+        if(!fs.exists(new Path(destPath))) {
+            throw new RuntimeException(destPath + " does not exist.");
+        }
+
+        FileUtil.copy(fs, new Path(srcPath), fs, new Path(destPath), false, new Configuration());
+    }
+
+    /**
      * Check the path is directory or not, the SourceType is used to find the file system
      * 
      * @param sourceFile
@@ -427,10 +484,11 @@ public class ShifuFileUtils {
      *            - destination file
      * @param sourceType
      *            - local/hdfs
+     * @return true if moving successfully, or false
      * @throws IOException
      *             - if any I/O exception in processing
      */
-    public static void move(String srcPath, String destPath, SourceType sourceType) throws IOException {
+    public static boolean move(String srcPath, String destPath, SourceType sourceType) throws IOException {
         if(StringUtils.isEmpty(srcPath) || StringUtils.isEmpty(destPath) || sourceType == null) {
             throw new IllegalArgumentException(String.format(
                     "Null or empty parameters srcDataPath:%s, dstDataPath:%s, sourceType:%s", srcPath, destPath,
@@ -443,7 +501,13 @@ public class ShifuFileUtils {
             // ignore delete failed, it's ok.
         }
 
-        fs.rename(new Path(srcPath), new Path(destPath));
+        if(fs.exists(new Path(srcPath))) {
+            // copy file only when source file exists.
+            fs.rename(new Path(srcPath), new Path(destPath));
+            return true;
+        }
+
+        return false;
     }
 
     /**
@@ -590,6 +654,11 @@ public class ShifuFileUtils {
     }
 
     public static FileStatus[] getFilePartStatus(String filePath, SourceType sourceType) throws IOException {
+        return getFilePartStatus(filePath, sourceType, Constants.HADOOP_PART_PREFIX);
+    }
+
+    public static FileStatus[] getFilePartStatus(String filePath, SourceType sourceType, final String partFilePrefix)
+            throws IOException {
         FileSystem fs = getFileSystemBySourceType(sourceType);
 
         PathFilter filter = new PathFilter() {
@@ -597,7 +666,7 @@ public class ShifuFileUtils {
             public boolean accept(Path path) {
                 // FIXME, should only skip _SUCCESS, .pig_header such files, not start from part, some files may not
                 // start from part.
-                return path.getName().startsWith("part");
+                return path.getName().startsWith(partFilePrefix);
             }
         };
 
@@ -610,11 +679,47 @@ public class ShifuFileUtils {
         }
 
         if(fileStatsArr == null || fileStatsArr.length == 0) {
-            // protected by reading glob status agaion
-            fileStatsArr = fs.globStatus(new Path(filePath), filter);
+            // protected by reading glob status again
+            fileStatsArr = fs.globStatus(new Path(filePath));
         }
 
         return fileStatsArr;
+    }
+
+    public static List<FileStatus> getFileStatus(String filePath, SourceType sourceType) throws IOException {
+        List<FileStatus> fileStatusList = new ArrayList<FileStatus>();
+        FileSystem fs = getFileSystemBySourceType(sourceType);
+        FileStatus[] fileStatusArr = fs.globStatus(new Path(filePath));
+        if ( fileStatusArr != null && fileStatusArr.length > 0 ) {
+            for ( FileStatus fileStatus : fileStatusArr ) {
+                fetchFileStatus(fileStatus.getPath(), sourceType, fileStatusList);
+            }
+        }
+        return fileStatusList;
+    }
+
+    private static void fetchFileStatus(Path filePath, SourceType sourceType, List<FileStatus> fileStatusList)
+            throws IOException {
+        FileSystem fs = getFileSystemBySourceType(sourceType);
+        try {
+            FileStatus[] fileStatsArr = fs.listStatus(filePath);
+            for(FileStatus fileStatus : fileStatsArr) {
+                if(fileStatus.isDirectory()) {
+                    fetchFileStatus(fileStatus.getPath(), sourceType, fileStatusList);
+                } else {
+                    if(!isHiddenFile(fileStatus.getPath().getName())) {
+                        fileStatusList.add(fileStatus);
+                    }
+                }
+            }
+        } catch (IOException e) {
+            log.error("Fail to fetch file status for - {}", filePath.toString());
+            throw e;
+        }
+    }
+
+    public static boolean isHiddenFile(String fileName) {
+        return StringUtils.isBlank(fileName) || fileName.startsWith(".") || fileName.startsWith("_");
     }
 
     public static int getFilePartCount(String filePath, SourceType sourceType) throws IOException {
@@ -627,7 +732,7 @@ public class ShifuFileUtils {
 
         boolean isGzip = true;
         for(FileStatus fileStatus: fileStatsArr) {
-            if(!fileStatus.getPath().toString().endsWith("gz") && !fileStatus.getPath().toString().endsWith("gz")) {
+            if(!fileStatus.getPath().toString().endsWith("gz")) {
                 isGzip = false;
             }
         }
@@ -644,4 +749,53 @@ public class ShifuFileUtils {
         return size;
     }
 
+    public static boolean isCompressedFileOrDirectory(String filePath, SourceType sourceType) throws IOException {
+        boolean isCompressedFile = false;
+
+        FileStatus[] fileStatsArr = getFilePartStatus(filePath, sourceType);
+        for(FileStatus fileStatus: fileStatsArr) {
+            if(fileStatus.getPath().toString().endsWith("gz") || fileStatus.getPath().toString().endsWith("bz2")) {
+                isCompressedFile = true;
+                break;
+            }
+        }
+        return isCompressedFile;
+    }
+
+    public static void writeLines(@SuppressWarnings("rawtypes") Collection collection, String filePath,
+            SourceType sourceType) throws IOException {
+        BufferedWriter writer = getWriter(filePath, sourceType);
+        try {
+            for(Object object: collection) {
+                if(object != null) {
+                    writer.write(object.toString());
+                    writer.newLine();
+                }
+            }
+        } finally {
+            IOUtils.closeQuietly(writer);
+        }
+    }
+
+    public static void copyToLocal(SourceFile sourceFile, String localOutputPath) throws IOException {
+        copyToLocal(sourceFile, Constants.HADOOP_PART_PREFIX, localOutputPath);
+    }
+
+    public static void copyToLocal(SourceFile sourceFile, String partFilePrefix, String localOutputPath)
+            throws IOException {
+        HdfsPartFile hdfsPartFile = new HdfsPartFile(sourceFile.getPath(), sourceFile.getSourceType(), partFilePrefix);
+        BufferedWriter writer = new BufferedWriter(new FileWriter(localOutputPath));
+        String line = null;
+        try {
+            while((line = hdfsPartFile.readLine()) != null) {
+                writer.write(line);
+                writer.newLine();
+            }
+        } catch (Exception e) {
+            // ignore
+        } finally {
+            IOUtils.closeQuietly(writer);
+            hdfsPartFile.close();
+        }
+    }
 }
